feat(main): use either 'maskable' or 'any' as a purpose
It's not recommended to use the maskable and any purposes together anymore. This setting also raises
issues with the new lighthouse audit. With this fix, when maskable is provided as an option to the
library, it will be only setting maskable as a purpose in the manifest.

BREAKING CHANGE: Instead of setting 'maskable any' as a purpose, the library will be setting
'maskable' value only when provided the maskable option.

fix #679