fix(main): prevent unnecessary newlines on index HTML file
Introduce an opinionated simple formatting to the index HTML file when preferred to inject the
output automatically.

BREAKING CHANGE: The newlines created when re-saving meta tag changes on the same index HTML file
will be removed. The prettier step to re-format after library execution is now optional and nice to
have.

fix #434