chore(core): use `import type` and `node:` prefix
TypeScript allows specifying a type keyword on an import to indicate that the import doesn’t exist at runtime. This allows compilers to safely drop imports of types without looking for their definition. This also ensures that some modules are not loaded at runtime.

Using the node: protocol for Node.js builtin modules makes it more clear that its a builtin module and prevents it from ever being overridden with userland module, which may pose a security threat.