chore(core): refactor spreads in accumulators
Using spread in accumulators like .reduce() may result in degraded
performance as you create new object with every iteration. This is
rarely needed. This refactors the code to avoid spread in accumulators
and while at it, removes a TODO that will never resolve for the reasons
stated in the discussion in the link I removed.