chore(test): disable globals
It seems like you've decided to use Vitest globals function in ONE place only, which led to ESLint config longer than it needs to be, so I just went ahead and disabled that option and disallowed Vitest globals. Additionally, transformMode does not exist in Vitest, so it had no effect, so I removed it as well.